requirements.txt
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
pytest>=7.0.0
streamlit>=1.28.0
openai-whisper>=20231117
pandas>=2.0.0
